+++ all-builds.nix
@@ -4,6 +4,7 @@
   }
 , local-self ? import ./. self-args
 , supportedSystems ? [ builtins.currentSystem ]
+, __useNewerCompiler  ? false #true if one wants to use ghc 8.10.7
 }:
 
 let
@@ -41,7 +42,7 @@ let
     else [];
 
   perPlatform = lib.genAttrs cacheBuildSystems (system: let
-    reflex-platform = import ./dep/reflex-platform { inherit system; };
+    reflex-platform = import ./dep/reflex-platform { inherit system __useNewerCompiler; };
 
     mkPerProfiling = profiling: let
       obelisk = import ./. (self-args // { inherit system profiling; });