+++ release.nix
@@ -1,5 +1,4 @@
-{
-  supportedSystems ? [ "x86_64-linux" "x86_64-darwin" ]
+{ supportedSystems ? [ "x86_64-linux" ]
 }:
 rec {
   recurseForDerivations = true;